topk_large_indices: one kernel source per RISC role (reader/compute/writer)
The single-codepath consolidation: 7 kernel sources collapse to 3 -- one
reader, one compute, one writer -- with the tree roles selected by factory
defines (TOPK_TREE for rectangle members, +TOPK_TREE_ROOT for the root),
exactly the pattern the fused-body selection already uses. Node and root
tree bodies merged into one arm on constexpr role booleans (all divergent
flags constant-fold); shared helpers (copy_chunk, classic finish, neginf
marker, index materialization) moved once into the common header; writer
keeps two whole-function role bodies where the loop structures genuinely
differ. The two factories remain separate functions (flat work split vs
rectangle/tree schedule share little) but now share CB/reader helpers.

No instruction sequence, CB protocol, or merge/rebuild order changed:
per-role projections of the unified sources diff clean (whitespace/comment
normalized) against the previous per-role files, modulo constant-foldable
substitutions. Net -197 LOC, 4 files deleted.
