+++ ttnn/cpp/ttnn/operations/experimental/topk_large_indices/device/kernels/compute.cpp
@@ -2,13 +2,63 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
+// Unified topk_large_indices compute kernel. The factory selects the role per
+// CreateKernel site via compile-time defines:
+//
+//   (no role define) -- ROW-PARALLEL: each core reduces full rows and
+//   materializes the index output itself. The body mode is picked by a second
+//   define set from shape: FUSED_E2E (stay fused through every merge/rebuild,
+//   one global split per row), FUSED_SEGMENTED (per-<=32-chunk-segment fusion
+//   folded into an unfused cross-segment survivor), or neither (classic
+//   unfused per-chunk split, with the data-dependent chunk-skip early-out).
+//
+//   TOPK_TREE -- COLUMN-PARALLEL tree node (every rectangle core except the
+//   root). Phase 1 (leaf): reduce this core's chunk slice to one K survivor in
+//   DST. Phase 2 (tree): for each of this core's num_merges winning levels,
+//   consume one partner sequence from the recv CB (delivered by this core's
+//   writer via the pairwise handshake), copy it to DST operand-1, and
+//   merge+rebuild IN PLACE -- the survivor never leaves DST between levels.
+//   Phase 3 (ship): pack the survivor raw (FP32 values + UINT32 indices) for
+//   the writer to send to this core's winner.
+//
+//   TOPK_TREE + TOPK_TREE_ROOT -- COLUMN-PARALLEL tree root (rectangle core
+//   (0,0), slice 0). Same leaf reduction and in-DST tree merges, but the root
+//   never ships: every rebuild stays DESCENDING (it is always the next merge's
+//   operand 0), and after the last level it materializes the indices exactly
+//   like the row-parallel epilogue.
+//
+// Tree directions: the bitonic halver needs operand0 descending and operand1
+// ascending, so every DST-mutating action rebuilds DESCENDING except a node's
+// LAST one before shipping, which rebuilds ASCENDING (validated upstream by
+// test_topk_xl_rebuild_ascending: rebuild(asc) after a merge is the exact
+// rank-mirror of rebuild(desc)).
+//
+// Tree empty slice (valid_length cut, nodes only -- slice 0 always holds at
+// least one element):
+//   * num_merges == 0 -> nothing to compute; the writer ships the prefilled
+//     all--inf scratch sequence instead.
+//   * num_merges > 0 -> the FIRST partner is ADOPTED (copied to operand-0 and
+//     rebuilt) instead of merged; an ascending monotonic run is bitonic, so
+//     the rebuild is valid in either direction.
+
 #include <cstdint>
 #include "api/compute/compute_kernel_api.h"
 #include "api/compute/compute_kernel_hw_startup.h"
-#include "api/compute/pack_untilize.h"
-#include "api/compute/experimental/topk_xl.h"
-#include "api/compute/transpose_dest.h"
 #include "api/dataflow/circular_buffer.h"
+
+#include "topk_large_indices_compute_common.hpp"
+
+#ifdef TOPK_TREE
+#include "api/compute/tile_move_copy.h"  // copy_tile of received sequences
+#ifndef TOPK_TREE_ROOT
+#include "api/compute/pack.h"  // pack_block of the shipped survivor
+#endif
+#endif
+#if !defined(TOPK_TREE) || defined(TOPK_TREE_ROOT)
+#include "api/compute/pack_untilize.h"  // rank-order index materialization
+#endif
+
+#ifndef TOPK_TREE
 
 #include "topk_large_indices_chunk_skip.hpp"
 
@@ -26,112 +76,27 @@ constexpr bool kChunkSkipEnable = true;
 #include "api/debug/dprint.h"
 #endif
 
-#ifdef TRISC_MATH
-namespace ckernel::sfpu {
-
-// topk_large_indices keeps final values and indices in the TopK XL LLK DST
-// layout until the last rank-order materialization step. The generic compute
-// APIs (`isneginf_tile` + `where_tile`) operate on normal tile layouts and do
-// not line up with this intermediate value/index pairing, so they only replaced
-// a subset of the final -inf lanes during validation.
-//
-// Keep this as op-local SFPU functionality for now instead of exporting a
-// public LLK API: it is tied to the final TopK XL LLK DST contract below,
-// where the value words start at the normal `idst` base and the UINT32 index
-// words start at `indices_offset`. The helper walks that layout directly,
-// compares final values against exact BF16 -inf stored in the FP32 DST
-// container (`0xFF800000`), and conditionally writes the sentinel index
-// `0xFFFFFFFF`.
-inline void _topk_large_indices_mark_neginf_indices_init_() {
-    addr_mod_t{
-        .srca = {.incr = 0},
-        .srcb = {.incr = 0},
-        .dest = {.incr = 2},
-    }
-        .set(ADDR_MOD_0);
-}
-
-template <uint32_t K>
-inline void _topk_large_indices_mark_neginf_indices_() {
-    static_assert(K == 512 || K == 1024 || K == 2048, "K must be 512, 1024, or 2048");
-    constexpr uint32_t tiles_per_sequence = K == 2048 ? 2 : 1;
-    constexpr uint32_t indices_offset = tiles_per_sequence * 64;
-    constexpr uint32_t iterations = (K == 512 ? 1 : K == 1024 ? 2 : 4) * 16;
-
-    TTI_SFPLOADI(p_sfpu::LREG2, sfpi::SFPLOADI_MOD0_LOWER, 0x0000);
-    TTI_SFPLOADI(p_sfpu::LREG2, sfpi::SFPLOADI_MOD0_UPPER, 0xFF80);
-    TTI_SFPLOADI(p_sfpu::LREG3, sfpi::SFPLOADI_MOD0_LOWER, 0xFFFF);
-    TTI_SFPLOADI(p_sfpu::LREG3, sfpi::SFPLOADI_MOD0_UPPER, 0xFFFF);
-
-    for (uint32_t i = 0; i < iterations; ++i) {
-        TTI_SFPLOAD(p_sfpu::LREG0, InstrModLoadStore::INT32, ADDR_MOD_7, 0);
-        TTI_SFPXOR(0, p_sfpu::LREG2, p_sfpu::LREG0, 0);
-        TTI_SFPSETCC(0, p_sfpu::LREG0, 0, sfpi::SFPSETCC_MOD1_LREG_EQ0);
-        TTI_SFPSTORE(p_sfpu::LREG3, InstrModLoadStore::INT32, ADDR_MOD_0, indices_offset);
-        TTI_SFPENCC(0, 0, 0, 0);
-    }
-}
-
-}  // namespace ckernel::sfpu
-#endif
-
 namespace {
 
-constexpr uint32_t elements_per_tile = TILE_R_DIM * TILE_C_DIM;
+using topk_large_indices::copy_chunk;
+using topk_large_indices::finish_chunk_classic;
 
-template <uint32_t K>
-FORCE_INLINE void materialize_index_rank_order(uint32_t idst, uint32_t indices_cb) {
-    static_assert(K == 512 || K == 1024 || K == 2048, "K must be 512, 1024, or 2048");
-    constexpr uint32_t tiles_per_sequence = (K + elements_per_tile - 1) / elements_per_tile;
-
-    transpose_dest_init<true, false>(indices_cb);
-    for (uint32_t t = 0; t < tiles_per_sequence; ++t) {
-        transpose_dest<true, false>(idst + tiles_per_sequence + t);
-    }
-}
-
-template <uint32_t K>
-FORCE_INLINE void mark_neginf_indices(uint32_t idst) {
-    static_assert(K == 512 || K == 1024 || K == 2048, "K must be 512, 1024, or 2048");
-    MATH((ckernel::sfpu::_topk_large_indices_mark_neginf_indices_init_()));
-    MATH((_llk_math_eltwise_unary_sfpu_params_(
-        ckernel::sfpu::_topk_large_indices_mark_neginf_indices_<K>, idst, VectorMode::RC_custom)));
-}
-
-// First half of chunk processing: pull the chunk from the input CB into DST.
-// Always runs -- the chunk must be resident in DST to be inspected at all.
-template <uint32_t K>
-FORCE_INLINE void copy_chunk(CircularBuffer& input_cb, uint32_t dst_base, uint32_t active_elements) {
-    constexpr uint32_t tiles_per_sequence = (K + elements_per_tile - 1) / elements_per_tile;
-    const uint32_t input_cb_id = input_cb.get_cb_id();
-    input_cb.wait_front(tiles_per_sequence);
-    topk_xl_copy_tile_init(input_cb_id);
-    topk_xl_copy_tile<K>(input_cb_id, dst_base, 0, active_elements);
-    input_cb.pop_front(tiles_per_sequence);
-}
-
-// Second half: stamp fused LSB indices and locally sort. FUSED_E2E rows stamp
-// the RUNTIME chunk id into index bits [15:11] and stay fused through every
-// merge/rebuild (one global split per row at the end); the classic path splits
-// to unfused values + row-major indices per chunk and advances the chunk base.
+// Second half of chunk processing: stamp fused LSB indices and locally sort.
+// FUSED_E2E rows stamp the RUNTIME chunk id into index bits [15:11] and stay
+// fused through every merge/rebuild (one global split per row at the end); the
+// classic path splits to unfused values + row-major indices per chunk and
+// advances the chunk base.
 template <uint32_t K>
 FORCE_INLINE void finish_chunk(uint32_t dst_base, bool ascending, uint32_t chunk_id) {
-    topk_xl_add_lsb_indices_init();
 #ifdef FUSED_E2E
+    topk_xl_add_lsb_indices_init();
     topk_xl_add_lsb_indices_rt<K>(dst_base, chunk_id);
 
     topk_xl_init<K, true>();
     topk_xl_local_sort<K>(dst_base, ascending);
 #else
     (void)chunk_id;
-    topk_xl_add_lsb_indices<K, 0>(dst_base);
-
-    topk_xl_init<K, true>();
-    topk_xl_local_sort<K>(dst_base, ascending);
-
-    topk_xl_separate_indices_row_major_reinit();
-    topk_xl_separate_indices_row_major<K>(dst_base);
-    topk_xl_separate_indices_row_major_advance_chunk_base<K>();
+    finish_chunk_classic<K>(dst_base, ascending);
 #endif
 }
 
@@ -160,20 +125,174 @@ FORCE_INLINE void process_chunk(
 
 }  // namespace
 
+#endif  // !TOPK_TREE
+
 void kernel_main() {
+    using namespace topk_large_indices;
+
     const uint32_t num_rows = get_arg_val<uint32_t>(0);
     const uint32_t num_chunks = get_arg_val<uint32_t>(1);
     const uint32_t tail_elements = get_arg_val<uint32_t>(2);
 
+#if defined(TOPK_TREE_ROOT)
+    constexpr uint32_t input_cb = get_compile_time_arg_val(0);
+    constexpr uint32_t indices_out_cb = get_compile_time_arg_val(1);
+    constexpr uint32_t recv_cb = get_compile_time_arg_val(2);
+    constexpr uint32_t K = get_compile_time_arg_val(3);
+#elif defined(TOPK_TREE)
+    constexpr uint32_t input_cb = get_compile_time_arg_val(0);
+    constexpr uint32_t ship_values_cb = get_compile_time_arg_val(1);
+    constexpr uint32_t ship_indices_cb = get_compile_time_arg_val(2);
+    constexpr uint32_t recv_cb = get_compile_time_arg_val(3);
+    constexpr uint32_t K = get_compile_time_arg_val(4);
+#else
     constexpr uint32_t input_cb = get_compile_time_arg_val(0);
     constexpr uint32_t indices_cb = get_compile_time_arg_val(1);
     constexpr uint32_t K = get_compile_time_arg_val(2);
     constexpr uint32_t USER_K = get_compile_time_arg_val(3);
+#endif
 
     static_assert(K == 512 || K == 1024 || K == 2048, "K must be 512, 1024, or 2048");
     constexpr uint32_t tiles_per_sequence = (K + elements_per_tile - 1) / elements_per_tile;
     constexpr uint32_t sequence_tiles = tiles_per_sequence * 2u;
     constexpr uint32_t slot0 = 0;
+
+#ifdef TOPK_TREE
+    // ------------------------------------------------------------------
+    // COLUMN-PARALLEL tree body (node and root)
+    // ------------------------------------------------------------------
+    const uint32_t start_chunk = get_arg_val<uint32_t>(3);  // always 0 for slice 0; kept for arg-layout parity
+    const uint32_t num_merges = get_arg_val<uint32_t>(4);
+
+#ifdef TOPK_TREE_ROOT
+    constexpr bool is_tree_root = true;
+#else
+    constexpr bool is_tree_root = false;
+#endif
+    constexpr uint32_t slot1 = sequence_tiles;
+
+    // An empty slice (valid_length cut) with no partners contributes nothing;
+    // the writer ships the -inf scratch sequence. The root's slice 0 always
+    // contains at least one element (valid_length >= 1), so it never takes
+    // the empty/adopt paths and they fold away.
+    const bool own_empty = !is_tree_root && (num_chunks == 0);
+    if (own_empty && num_merges == 0) {
+        return;
+    }
+
+#ifdef TOPK_TREE_ROOT
+    compute_kernel_hw_startup(input_cb, indices_out_cb);
+#else
+    compute_kernel_hw_startup(input_cb, ship_values_cb);
+#endif
+
+    CircularBuffer input_cb_obj(input_cb);
+    CircularBuffer recv_obj(recv_cb);
+#ifdef TOPK_TREE_ROOT
+    CircularBuffer indices_out_obj(indices_out_cb);
+#else
+    CircularBuffer ship_values_obj(ship_values_cb);
+    CircularBuffer ship_indices_obj(ship_indices_cb);
+#endif
+
+    for (uint32_t row = 0; row < num_rows; ++row) {
+        tile_regs_acquire();
+
+        if (!own_empty) {
+            // Leaf reduction of this core's chunk slice. The chunk base is
+            // latched at 0 and stepped with the all-constant advance
+            // primitive: a runtime chunk_base cannot reach the
+            // TTI-immediate config write inside the init.
+            topk_xl_separate_indices_row_major_init_static<0, 0>();
+            for (uint32_t c = 0; c < start_chunk; ++c) {
+                topk_xl_separate_indices_row_major_advance_chunk_base<K>();
+            }
+
+            // A node whose leaf survivor ships directly (no tree merges)
+            // rebuilds ASCENDING on its last DST action; the root (and any
+            // survivor with merges still ahead) rebuilds DESCENDING.
+            const bool ship_after_leaf = !is_tree_root && (num_merges == 0);
+
+            const uint32_t first_chunk_elements = (num_chunks == 1) ? tail_elements : K;
+            process_chunk<K>(input_cb_obj, slot0, first_chunk_elements, false);
+
+            if (num_chunks == 1) {
+                topk_xl_init<K, false>();
+                topk_xl_rebuild<K, false>(slot0, ship_after_leaf);
+            }
+
+            for (uint32_t chunk = 1; chunk < num_chunks; ++chunk) {
+                const uint32_t active_elements = (chunk + 1 == num_chunks) ? tail_elements : K;
+                process_chunk<K>(input_cb_obj, slot1, active_elements, true);
+
+                topk_xl_init<K, false>();
+                topk_xl_merge<K, false>(slot0);
+                const bool last_chunk = (chunk + 1 == num_chunks);
+                topk_xl_rebuild<K, false>(slot0, last_chunk && ship_after_leaf);
+            }
+        }
+
+        // Tree merges: one partner sequence per winning level, in level order.
+        for (uint32_t m = 0; m < num_merges; ++m) {
+            const bool ship_next = !is_tree_root && (m + 1 == num_merges);
+            const bool adopt = own_empty && (m == 0);
+            const uint32_t dst_base = adopt ? slot0 : slot1;
+
+            recv_obj.wait_front(sequence_tiles);
+            // EXPLICIT format reconfig (bf16 -> UInt32): the preceding chunk phase
+            // (topk_xl_copy_tile_init) hw-configures the unpacker for the BF16
+            // input CB, and the plain _short init's state_configure is a no-op
+            // in production JIT builds (the compute-kernel sentinel is compiled
+            // out), so the _with_dt variant's llk reconfig is required here.
+            // The reverse transition is explicit inside topk_xl_copy_tile_init.
+            copy_tile_to_dst_init_short_with_dt(input_cb, recv_cb);
+            for (uint32_t t = 0; t < sequence_tiles; ++t) {
+                copy_tile(recv_cb, t, dst_base + t);
+            }
+            recv_obj.pop_front(sequence_tiles);
+
+            topk_xl_init<K, false>();
+            if (!adopt) {
+                topk_xl_merge<K, false>(slot0);
+            }
+            topk_xl_rebuild<K, false>(slot0, ship_next /* ascending for the winner's halver */);
+        }
+
+#ifdef TOPK_TREE_ROOT
+        // Root epilogue: materialize the indices exactly like the row-parallel
+        // kernel (mark_neginf + rank-order transpose + pack_untilize).
+        mark_neginf_indices<K>(slot0);
+        materialize_index_rank_order<K>(slot0, indices_out_cb);
+#endif
+
+        tile_regs_commit();
+        tile_regs_wait();
+
+#ifdef TOPK_TREE_ROOT
+        indices_out_obj.reserve_back(1);
+        pack_untilize_dest_init<tiles_per_sequence, tiles_per_sequence>(indices_out_cb);
+        pack_untilize_dest<tiles_per_sequence, tiles_per_sequence>(indices_out_cb, 1, 0, slot0 + tiles_per_sequence);
+        indices_out_obj.push_back(1);
+#else
+        // Ship the survivor raw: bit-exact FP32 value tiles then UINT32 index
+        // tiles; the winner's copy_tile round-trips the DST image unchanged.
+        ship_values_obj.reserve_back(tiles_per_sequence);
+        pack_block(slot0, ship_values_cb, tiles_per_sequence);
+        ship_values_obj.push_back(tiles_per_sequence);
+
+        ship_indices_obj.reserve_back(tiles_per_sequence);
+        pack_block(slot0 + tiles_per_sequence, ship_indices_cb, tiles_per_sequence);
+        ship_indices_obj.push_back(tiles_per_sequence);
+#endif
+
+        tile_regs_release();
+    }
+
+#else  // !TOPK_TREE
+
+    // ------------------------------------------------------------------
+    // ROW-PARALLEL body (FUSED_SEGMENTED / FUSED_E2E / classic)
+    // ------------------------------------------------------------------
 #ifdef FUSED_SEGMENTED
     // Cross-segment survivor (unfused [v,i]) lives at tiles 0..2*tps-1; the
     // in-flight fused segment at 2*tps..3*tps-1; its incoming chunk after it.
@@ -355,4 +474,6 @@ void kernel_main() {
 
         tile_regs_release();
     }
+
+#endif  // TOPK_TREE
 }
