--- ttnn/cpp/ttnn/operations/experimental/topk_large_indices/device/kernels/reader_local.cpp
@@ -1,62 +1,0 @@
-// SPDX-FileCopyrightText: © 2026 Tenstorrent AI ULC
-//
-// SPDX-License-Identifier: Apache-2.0
-
-// Column-parallel local reader: identical to reader.cpp except each core reads
-// only its contiguous slice of every row, offset by slice_offset_bytes.
-// Kept as a separate kernel (instead of an extra runtime arg on reader.cpp) so
-// the row-parallel path's reader source stays byte-identical — its JIT cache
-// entry, binary, and perf baselines are untouched by the column-parallel work.
-
-#include "api/dataflow/circular_buffer.h"
-#include "api/dataflow/dataflow_api.h"
-#include "api/dataflow/noc.h"
-#include "api/tensor/noc_traits.h"
-
-void kernel_main() {
-    const uint32_t src_addr = get_arg_val<uint32_t>(0);
-    const uint32_t start_row = get_arg_val<uint32_t>(1);
-    const uint32_t num_rows = get_arg_val<uint32_t>(2);
-    const uint32_t num_chunks = get_arg_val<uint32_t>(3);
-    const uint32_t tail_chunk_bytes = get_arg_val<uint32_t>(4);
-    const uint32_t input_page_bytes = get_arg_val<uint32_t>(5);
-    // Byte offset of this core's slice within each row. An empty slice
-    // (num_chunks == 0, valid_length cut) reads nothing.
-    const uint32_t slice_offset_bytes = get_arg_val<uint32_t>(6);
-
-    constexpr uint32_t cb_in = get_compile_time_arg_val(0);
-    constexpr uint32_t chunk_bytes = get_compile_time_arg_val(1);
-    constexpr uint32_t tile_bytes = get_compile_time_arg_val(2);
-    constexpr uint32_t tiles_per_chunk = get_compile_time_arg_val(3);
-    constexpr auto input_args = TensorAccessorArgs<4>();
-
-    const auto input = TensorAccessor(input_args, src_addr, input_page_bytes);
-    CircularBuffer input_cb(cb_in);
-    Noc noc;
-
-    for (uint32_t local_row = 0; local_row < num_rows; ++local_row) {
-        const uint32_t row = start_row + local_row;
-        for (uint32_t chunk = 0; chunk < num_chunks; ++chunk) {
-            const uint32_t active_chunk_bytes = (chunk + 1 == num_chunks) ? tail_chunk_bytes : chunk_bytes;
-            input_cb.reserve_back(tiles_per_chunk);
-            for (uint32_t tile = 0; tile < tiles_per_chunk; ++tile) {
-                const uint32_t tile_offset = tile * tile_bytes;
-                const uint32_t read_bytes =
-                    tile_offset < active_chunk_bytes
-                        ? (active_chunk_bytes - tile_offset < tile_bytes ? active_chunk_bytes - tile_offset
-                                                                         : tile_bytes)
-                        : 0;
-                if (read_bytes != 0) {
-                    noc.async_read(
-                        input,
-                        input_cb,
-                        read_bytes,
-                        {.page_id = row, .offset_bytes = slice_offset_bytes + chunk * chunk_bytes + tile_offset},
-                        {.offset_bytes = tile_offset});
-                }
-            }
-            noc.async_read_barrier();
-            input_cb.push_back(tiles_per_chunk);
-        }
-    }
-}
